refactor: 데이터베이스에만 있던 이름과 가격 제약사항을 애플리케이션 영역의 VO에도 구현, 관련 코드를 도메인 단위 테스트로 구현
